% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpmccr-package.R
\docType{package}
\name{bnpmccr-package}
\alias{bnpmccr}
\alias{bnpmccr-package}
\title{bnpmccr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YujinJeong0202/bnpmccr}
  \item Report bugs at \url{https://github.com/YujinJeong0202/bnpmccr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
