% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpmccr-package.R
\docType{package}
\name{bnpmccr-package}
\alias{bnpmccr}
\alias{bnpmccr-package}
\title{bnpmccr: Bayesian Nonparametric Conditional Copula Regression}
\description{
Implements Bayesian nonparametric multivariate conditional copula regression with varying coefficients for mixed-type responses. The package provides functions for fitting flexible marginal regressions and covariate-dependent Gaussian copula mixtures using Markov chain Monte Carlo. It is intended to support the reproducibility of the accompanying manuscript.
}
\author{
\strong{Maintainer}: Anonymous Author \email{anonymous@example.com}

}
\keyword{internal}
