% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpmccr-package.R
\docType{package}
\name{bnpmccr-package}
\alias{bnpmccr}
\alias{bnpmccr-package}
\title{bnpmccr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
 }

}
\author{
\strong{Maintainer}: Anonymous Author \email{anonymous@example.com}

}
\keyword{internal}
